Remove LSP handlers and inline the actual functionality
